реализация выбора таргета и изменение спавна мобов

+++ src/main/java/com/server/DataBase.java
@@ -1,9 +1,7 @@
 package com.server;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.Hashtable;
-import java.util.Set;
 
 public final class DataBase {
     final static ArrayList<Byte> VENDOR_ID = new ArrayList<>(3);
@@ -14,8 +12,8 @@ public final class DataBase {
         VENDOR_ID.add((byte) 0);
         VENDOR_ID.add((byte) 1);
         VENDOR_ID.add((byte) 2);
-        MOBS.put((byte)4, new Mob((byte) 4,1,1,1, CardTableType.MOB, 0, CardTableSubType.GOBLIN, 1));
-        MOBS.put((byte)0, new Mob((byte) 0,0,0,0, CardTableType.VENDOR, 9999, CardTableSubType.TRADER, 0));
+        MOBS.put((byte)4, new Mob((byte) 4,1, (byte) 1,1, CardTableType.MOB, 0, CardTableSubType.GOBLIN, 1));
+        MOBS.put((byte)0, new Mob((byte) 0,0, (byte) 0,0, CardTableType.VENDOR, 9999, CardTableSubType.TRADER, 0));
         ITEMS.put((byte)0, new Item(1,0, InventoryType.WEAPON, (byte) 0,999,(byte)0));
         ITEMS.put((byte)1, new Item(1,0, InventoryType.WEAPON, (byte) 0,999,(byte)0));
         ITEMS.put((byte)17, new Item(1,0, InventoryType.FOOD, (byte) 1,0,(byte)0));
@@ -25,14 +23,14 @@ public final class DataBase {
     static class Mob {
         private byte mID;
         private int mValueOne;
-        private int mValueTwo;
+        private byte mValueTwo;
         private int mMoney;
         private int mType;
         private int mGearScore;
         private int mSubType;
         private int mExperience;
 
-        public Mob(byte ID, int valueOne, int valueTwo, int money, int type, int gearScore, int subType, int experience) {
+        public Mob(byte ID, int valueOne, byte valueTwo, int money, int type, int gearScore, int subType, int experience) {
             mID = ID;
             mValueOne = valueOne;
             mValueTwo = valueTwo;
@@ -49,7 +47,7 @@ public final class DataBase {
         public int getValueOne() {
             return mValueOne;
         }
-        public int getValueTwo() {
+        public byte getValueTwo() {
             return mValueTwo;
         }
         public int getMoney() {
